Add support for generating sha256-based jigdo & template
This is format 2.0

Add choice of checksum algorithm on the command line, still using md5
as a default for backwards compatibility (and then still generating
format v1.2). If using sha256, then use the new template descriptor
types etc.

Lots of updates to make this happen, and lots of internal functions
modified to take both md5 and sha256 data rather than just md5, and an
extra parameter tpo say which is in use.